requirements.txt
# --------- pytorch --------- #
torch>=1.10.0
torchvision>=0.11.0
pytorch-lightning==1.6.4
torchmetrics==0.9.1

# --------- hydra --------- #
hydra-core==1.2.0
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
wandb==0.11.2

# --------- models --------- #
scikit-learn==1.1.3
einops==0.4.1
pytorch-tabnet==3.1.1
pytorch-widedeep==1.1.1
statsmodels==0.13.3
xgboost==1.7.0
catboost==1.1.1
lightgbm==3.3.3

# --------- xai --------- #
shap==0.39.0

# --------- figures --------- #
plotly==5.11.0
matplotlib==3.4.3
seaborn==0.12.1
kaleido==0.2.1

# --------- dataframes --------- #
pandas==1.5.1
openpyxl==3.0.10

# --------- imputation --------- #
impyute==0.0.8

# --------- others --------- #
python-dotenv==0.21.0   # loading env variables from .env file
rich==12.6.0            # beautiful text formatting in terminal
pyrootutils==1.0.4      # for getting root
python-slugify==8.0.1   # for valid filenames generation
